Feat:디테일 페이지 쿠폰 데이터 가져오기

+++ app/src/main/java/dev/kichan/marketplace/ui/MyApp.kt
@@ -52,7 +52,7 @@ fun MyApp(singlethone: SingleTonViewModel = SingleTonViewModel()) {
 
             composable("${Page.EventDetail.name}/{id}") {
                 it.arguments?.getString("id")?.let { id ->
-                    DetailPage(navController, id)
+                    DetailPage(navController, id.toLong())
                 }
             }
         }
